feature: support HTML directly within component tags
This resolves issues with content marked safe for rendering being incorrectly
escaped when used within a React component. Previously you had to workaround it
by using {% raw_html %} but was inconvenient, introduced another DOM node into
the tree, and not at all obvious when was required so was easy to miss.