feature: ViteBundler support custom 'unresolvedPath' option
'unresolvedPath' is a value added to the Vite manifest.json by a custom plugin. It's
used to resolve the used path (e.g. @internationalized/date) to the real entry point
for that package (e.g. "node_modules/@internationalized/date/dist/import.mjs").

The existing logic was simplistic and only handled looking for a matching index file
when a bare package was used. This worked for many cases, and always works locally,
but for packages imported from node_modules it isn't sufficient. This plugin maps
all usages in a project to the actual file to make this lookup easy.