fix: undiscoverable assets false positive fix
Fix cases where an asset may be treated as undiscoverable incorrectly. This occurs when
using a component as a var, e.g. {% component "package" "MyComponent" as my_var %}{% endcomponent %}
and then passing that as a prop to another component. When the prop is processed in the
component it is passed to it would previously throw an error about it being undiscoverable.
This fix accumulates the known resources as we go, so if the prop is encountered after the
node has been processed it is no longer considered unknown.

This would not have previously caused issues if the prop passed used an import path that was
used in the component (e.g. @alliancesoftware/ui components). With the new changes, this
would be much more common as using a new import from that package would be treated as a
unique usage.