fix: specify path_base when writing source
This helps in cases where the bundler `root_dir` has been changed and
no longer matches the default PROJECT_DIR. For example, the project source
may be in a sub directory within the repo, but the node_modules directory
or other shared source may be at a level higher. `root_dir` has to
encompass all source files and so must be set to the highest level.